Move LoginManager's encoder-decoder to example
- The LoginManager is base class for differnet auth types
- Currently only 2 types (base and oauth2) are defined in this
  repository
- The app (e.g. example app in this case) can add more types
- To do so moving implicit base encoder and decoder to example
